feat: Add StarRating component for displaying book ratings
This commit adds a new component called StarRating, which is used to display the rating of a book. The StarRating component takes a rating value as a prop and renders a star icon for each whole number value, and a half star icon for a decimal value. This component will be used in the book index and show pages to show the average rating of each book.